+++ packages/EasyApiPlatform/tests/Fixture/app/src/AdvancedSearchFilter/NameConverter/CustomConverter.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyApiPlatform\Tests\Fixture\App\AdvancedSearchFilter\Serializer\NameConverter;
+namespace EonX\EasyApiPlatform\Tests\Fixture\App\AdvancedSearchFilter\NameConverter;
 
 use Symfony\Component\Serializer\NameConverter\CamelCaseToSnakeCaseNameConverter;
