+++ packages/EasyApiPlatform/tests/Unit/src/Filter/AdvancedSearchFilterTest.php
@@ -11,7 +11,7 @@ use Doctrine\Persistence\ManagerRegistry;
 use EonX\EasyApiPlatform\Filter\AdvancedSearchFilter;
 use EonX\EasyApiPlatform\Tests\Fixture\App\AdvancedSearchFilter\ApiResource\Dummy;
 use EonX\EasyApiPlatform\Tests\Fixture\App\AdvancedSearchFilter\ApiResource\RelatedDummy;
-use EonX\EasyApiPlatform\Tests\Fixture\App\AdvancedSearchFilter\Serializer\NameConverter\CustomConverter;
+use EonX\EasyApiPlatform\Tests\Fixture\App\AdvancedSearchFilter\NameConverter\CustomConverter;
 use Prophecy\Argument;
 use Prophecy\Prophet;
